minor fix in PrintQ

+++ q-agent.cc
@@ -47,7 +47,7 @@ Action QAgent::GetNextMove(State state) {
 }
 
 void QAgent::PrintQ() {
-    unsigned int state_space{Q.size()},
+    unsigned int state_space{static_cast<unsigned int>(Q.size())},
                  action_space{Q[0].size()};
         
     std::cout << "Printing Q Matrix Values " << std::endl;
